refactor: move errorMsg dictionary to utils dir to be more accesible
