+++ tests/routes/patientTemplate/renamePatientTemplate.test.js
@@ -1,6 +1,6 @@
 const axios = require('axios')
 const { BASE_URL, getAuthToken } = require('../../jest.setup')
-const COMMON_MSG = require('../../errorMsg')
+const COMMON_MSG = require('../../../utils/errorMsg')
 const {
   createTestDoctor,
   deleteUser,
