+++ utils/errorMsg.js
@@ -1,16 +1,16 @@
 const COMMON_MSG = {
-  // 200
+  // 200: Success
   REQUEST_SUCCESS: 'Request succesful.',
-  // 400
+  // 400: Missing data on request
   MISSING_FIELDS: 'Missing Fields.',
-  // 403
+  // 403: Not authorized
   DOCTOR_IS_NOT_OWNER: 'Doctor is not the owner of template.',
-  // 404
+  // 404: Resource not found
   DOCTOR_NOT_FOUND: 'Doctor not found.',
   TEMPLATE_NOT_FOUND: 'Template not found.',
-  // 406
+  // 406: Resource already in use
   RECORDS_USING: 'Item with that id/name already exist',
-  // 409
+  // 409: Resources depend on this resource
   OPERATION_REJECTED: 'Could not modify item since other resources depend on it'
 }
